+++ account-kit/signer/src/index.ts
@@ -7,15 +7,16 @@ export {
 } from "./client/index.js";
 export type * from "./client/types.js";
 export {
+  NotAuthenticatedError,
+  OAuthProvidersError,
+  MfaRequiredError,
+} from "./errors.js";
+export {
   DEFAULT_SESSION_MS,
   SessionManagerParamsSchema,
 } from "./session/manager.js";
 export type * from "./signer.js";
 export { AlchemyWebSigner } from "./signer.js";
+export type * from "./solanaSigner.js";
 export type * from "./types.js";
 export { AlchemySignerStatus } from "./types.js";
-export {
-  OAuthProvidersError,
-  NotAuthenticatedError,
-  MfaRequiredError,
-} from "./errors.js";