+++ account-kit/signer/src/session/manager.ts
@@ -144,6 +144,7 @@ export class SessionManager {
 
   public clearSession = () => {
     this.store.setState({ session: null });
+    localStorage.removeItem(`${this.sessionKey}:temporary`);
 
     if (this.clearSessionHandle) {
       clearTimeout(this.clearSessionHandle);
@@ -193,6 +194,7 @@ export class SessionManager {
      * We should revisit this later
      */
     if (session.expirationDateMs < Date.now()) {
+      this.client.disconnect();
       this.clearSession();
       return null;
     }
@@ -323,6 +325,7 @@ export class SessionManager {
     }
 
     this.clearSessionHandle = setTimeout(() => {
+      this.client.disconnect();
       this.clearSession();
     }, Math.min(session.expirationDateMs - Date.now(), Math.pow(2, 31) - 1));
   };