+++ examples/ui-demo/src/components/small-cards/Wrapper.tsx
@@ -1,5 +1,6 @@
 import { useConfigStore } from "@/state";
 import { MintCard7702, MintCardDefault } from "./MintCard";
+import { SolanaCard } from "./SolanaCard";
 import {
   TransactionsCard7702,
   TransactionsCardDefault,
@@ -14,12 +15,14 @@ export const SmallCardsWrapper = () => {
       {accountMode === "default" ? (
         <>
           <MintCardDefault />
+          <SolanaCard />
           <TransactionsCardDefault />
           <MFACard />
         </>
       ) : (
         <>
           <MintCard7702 />
+          <SolanaCard />
           <TransactionsCard7702 />
           <MFACard />
         </>